Refactor attester config in identity
