+++ identity-service/src/routes/healthCheck.js
@@ -15,7 +15,7 @@ const {
 
 const axios = require('axios')
 const moment = require('moment')
-const { REDIS_ATTEST_HEALTH_KEY } = require('../app.js')
+const { REDIS_ATTEST_HEALTH_KEY } = require('../utils/configureAttester')
 
 // Defaults used in relay health check endpoint
 const RELAY_HEALTH_TEN_MINS_AGO_BLOCKS = 120 // 1 block/5sec = 120 blocks/10 minutes
